actions/src/py_wand/requirements.txt
Wand
python-swiftclient
